move get_file_name to avoid circuliar import
